dumper: Prefer u64 for pointer types

+++ dumper/src/containers.rs
@@ -121,7 +121,7 @@ impl<C: Mem + Clone + NameTrait + StructsTrait + VersionTrait> Ptr<FName, C> {
         if mem.ue_version() < (4, 22) {
             // wtf :skull_emoji:
             let chunks = self
-                .map(|_| mem.fnamepool().0)
+                .map(|_| mem.fnamepool().0 as u64)
                 .cast::<Ptr<Ptr<Ptr<(), C>, C>, C>>()
                 .read()?;
 
@@ -167,7 +167,9 @@ impl<C: Mem + Clone + NameTrait + StructsTrait + VersionTrait> Ptr<FName, C> {
             });
         }
 
-        let blocks = self.map(|_| mem.fnamepool().0 + 0x10).cast::<Ptr<u8, C>>();
+        let blocks = self
+            .map(|_| mem.fnamepool().0 as u64 + 0x10)
+            .cast::<Ptr<u8, C>>();
 
         let block_index = (value >> 16) as usize;
         let offset = if case_preserving {
